+++ Assets/OxGKit/Utilities/Scripts/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## CHANGELOG
 
+## [1.1.0] - 2024-10-21
+- Added CursorManager support for both static and dynamic modes, and allow setting the cursor's animation state.
+- Added CursorManager prefab in Samples.
+- Added CursorManager demo in Samples.
+
 ## [1.0.5] - 2024-06-20
 - Modified NewSingleton.
 - Modified MonoSingleton.